% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSconsole-package.R
\docType{package}
\name{JSconsole-package}
\alias{JSconsole-package}
\title{JSconsole addin}
\description{
This package provides a RStudio addin to send some JavaScript code in the
V8 console. To run the addin, open a JavaScript file, select some lines if
you want to send only this piece of code, and select \code{JSconsole} from
the Addins menu within RStudio.
}
